Assign the 'constructor' property to recover good console debug strings.
The javascript console appears to use the *name* of the *function* assigned
to the 'constructor' property of an object in order to generate the
brief string description of an object.  So assign to 'constructor' in our
constructor method -- by convention naming the __init__ method and using
that as the object's constructor function.  This allows us to have
pretty debug strings in the console w/o crufting our code up *too* much.