asm-llvm: parse/validate function table dereferences.
These are a bit tricky because they are all forward references, but we've
already done the hard part.  We just return a new TableBinding type which
lets us defer typechecking (and codegen) until we see the call arguments.
We can then infer a type for the function table and check it for consistency.